# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_oojoseph67_ecommerce_internal_dto.AddToCartRequest:
    properties:
      product_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      quantity:
        example: 2
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.AdminSignupRequest:
    properties:
      code:
        example: ADMIN001
        type: string
      email:
        example: admin@example.com
        type: string
      password:
        example: Admin123!
        type: string
    required:
    - code
    - email
    - password
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.UserResponse'
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.CartItemResponse:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      product:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ProductResponse'
      quantity:
        example: 2
        type: integer
      subtotal:
        example: 199.98
        type: number
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.CartResponse:
    properties:
      cart_item:
        items:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CartItemResponse'
        type: array
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      total:
        example: 199.98
        type: number
      total_cart_items:
        example: 2
        type: number
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.CategoryResponse:
    properties:
      description:
        example: Electronic devices and gadgets
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Electronics
        type: string
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: User123!
        type: string
    required:
    - email
    - password
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.LogoutRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.ProductImageResponse:
    properties:
      alt_text:
        example: Product image
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_primary:
        example: true
        type: boolean
      url:
        example: https://storage.example.com/products/image.jpg
        type: string
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CategoryResponse'
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        example: Premium wireless headphones with noise cancellation
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      images:
        items:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ProductImageResponse'
        type: array
      is_active:
        example: true
        type: boolean
      meta:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ResponseMeta'
      name:
        example: Wireless Headphones
        type: string
      price:
        example: 99.99
        type: number
      sku:
        example: WH-1000XM4
        type: string
      stock:
        example: 100
        type: integer
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.ResponseMeta:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.SignupRequest:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      password:
        example: User123!
        type: string
      phone:
        example: "+1234567890"
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - phone
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.UpdateCartItemRequest:
    properties:
      quantity:
        example: 3
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.UserResponse:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      last_name:
        example: Doe
        type: string
      phone:
        example: "+1234567890"
        type: string
      role:
        example: user
        type: string
    type: object
...
  github_com_oojoseph67_ecommerce_internal_utils_responses.Response:
    properties:
      data: {}
      error:
        example: invalid request
        type: string
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:1234
info:
  contact:
    email: oojoseph67@gmail.com
    name: Joseph
    url: https://josephbuilds.xyz
  description: A modern e-commerce API built with Go, Gin, and GORM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Commerce API
  version: "1"
paths:
  /auth/admin/login:
    post:
      consumes:
      - application/json
      description: Authenticate an admin account
      parameters:
      - description: Admin login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      summary: Admin login
      tags:
      - Auth
  /auth/admin/signup:
    post:
      consumes:
      - application/json
      description: Register a new admin account with a valid admin code
      parameters:
      - description: Admin signup credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.AdminSignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      summary: Admin registration
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user account
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      summary: User login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the refresh token
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      summary: User logout
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      summary: Refresh access token
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User signup credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      summary: User registration
      tags:
      - Auth
  /cart/:
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user's shopping cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CartResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Get user cart
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Add a product to the authenticated user's shopping cart
      parameters:
      - description: Product to add to cart
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CartResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/items/{id}:
    delete:
      consumes:
      - application/json
      description: Remove an item from the shopping cart
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
    patch:
      consumes:
      - application/json
      description: Update the quantity of an item in the shopping cart
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CartItemResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"